OCT-111: moved FindOneCatalogByIdQuery
