+++ components/catalogs/back/src/Application/Persistence/Catalog/FindOneCatalogByIdQueryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Application\Persistence;
+namespace Akeneo\Catalogs\Application\Persistence\Catalog;
 
 use Akeneo\Catalogs\ServiceAPI\Model\Catalog;
