+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/FindOneCatalogByIdQueryTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Catalog;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\FindOneCatalogByIdQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\FindOneCatalogByIdQuery;
 use Akeneo\Catalogs\ServiceAPI\Model\Catalog;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
@@ -12,7 +12,7 @@ use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\FindOneCatalogByIdQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Catalog\FindOneCatalogByIdQuery
  */
 class FindOneCatalogByIdQueryTest extends IntegrationTestCase
 {
